TST: Merge equation parsing into solution strategy
