--- tests/integration/models/test_equations.py
@@ -1,102 +1,0 @@
-"""Tests 'final' equations employed in the models.
-
-The tests here considers equations that are present in the global system to be solved,
-as opposed to simpler relations (typically constitutive relations) that are used to
-construct these global equations.
-
-See also: tests.integration.models.test_constitutive_laws.
-
-"""
-import pytest
-
-from . import setup_utils
-
-
-@pytest.mark.parametrize(
-    "model_type,equation_name,only_codimension",
-    [
-        ("mass_balance", "mass_balance_equation", None),
-        ("mass_balance", "interface_darcy_flux_equation", None),
-        ("momentum_balance", "momentum_balance_equation", 0),
-        ("momentum_balance", "interface_force_balance_equation", 1),
-        ("momentum_balance", "normal_fracture_deformation_equation", 1),
-        ("momentum_balance", "tangential_fracture_deformation_equation", 1),
-        ("energy_balance", "energy_balance_equation", None),
-        ("energy_balance", "interface_enthalpy_flux_equation", None),
-        ("energy_balance", "interface_fourier_flux_equation", None),
-        # Energy balance inherits mass balance equations. Test one of these as well.
-        ("energy_balance", "mass_balance_equation", None),
-        ("poromechanics", "mass_balance_equation", None),
-        ("poromechanics", "momentum_balance_equation", 0),
-        ("poromechanics", "interface_force_balance_equation", 1),
-        ("thermoporomechanics", "interface_fourier_flux_equation", None),
-    ],
-)
-# Run the test for models with and without fractures. We skip the case of more than one
-# fracture, since it seems unlikely this will uncover any errors that will not be found
-# with the simpler setups. Activate more fractures if needed in debugging.
-@pytest.mark.parametrize(
-    "num_fracs",
-    [  # Number of fractures
-        0,
-        1,
-        pytest.param(2, marks=pytest.mark.skipped),
-        pytest.param(3, marks=pytest.mark.skipped),
-    ],
-)
-@pytest.mark.parametrize("domain_dim", [2, 3])
-def test_parse_equations(
-    model_type, equation_name, only_codimension, num_fracs, domain_dim
-):
-    """Test that equation parsing works as expected.
-
-    Currently tested are the relevant equations in the mass balance and momentum balance
-    models. To add a new model, add a new class in setup_utils.py and expand the
-    test parameterization accordingly.
-
-    Parameters:
-        model_type: Type of model to test. Currently supported are "mass_balance" and
-            "momentum_balance". To add a new model, add a new class in setup_utils.py.
-        equation_name: Name of the method to test.
-        domain_inds: Indices of the domains for which the method should be called.
-            Some methods are only defined for a subset of domains.
-
-    """
-    if only_codimension is not None:
-        if only_codimension == 0 and num_fracs > 0:
-            # If the test is to be run on the top domain only, we need not consider
-            # models with fractures (note that since we iterate over num_fracs, the
-            # test will be run for num_fracs = 0, which is the top domain only)
-            return
-        elif only_codimension == 1 and num_fracs == 0:
-            # If the point of the test is to check the method on a fracture, we need
-            # not consider models without fractures.
-            return
-        else:
-            # We will run the test, but only on the specified codimension.
-            dimensions_to_assemble = domain_dim - only_codimension
-    else:
-        # Test on subdomains or interfaces of all dimensions
-        dimensions_to_assemble = None
-
-    if domain_dim == 2 and num_fracs > 2:
-        # The 2d models are not defined for more than two fractures.
-        return
-
-    # Set up an object of the prescribed model
-    setup = setup_utils.model(model_type, domain_dim, num_fracs=num_fracs)
-    # Fetch the relevant method of this model and extract the domains for which it is
-    # defined.
-    method = getattr(setup, equation_name)
-    domains = setup_utils.subdomains_or_interfaces_from_method_name(
-        setup.mdg, method, dimensions_to_assemble
-    )
-
-    # Call the discretization method.
-    setup.equation_system.discretize()
-
-    # Assemble the matrix and right hand side for the given equation. An error here will
-    # indicate that something is wrong with the way the conservation law combines
-    # terms and factors (e.g., grids, parameters, variables, other methods etc.) to form
-    # an Ad operator object.
-    setup.equation_system.assemble_subsystem({equation_name: domains})
